moved recipePage and refactored to new components
